Fix compile errors found by dotnet build; add test.ps1
Issues caught:
- TechSupport.Shared missing explicit System.IO.Pipelines package ref
- DxgiScreenCapture: Vortice 3.5 Texture2DDescription Width/Height and
  IDXGIOutputDuplication.AcquireNextFrame timeout are int, not uint
- ConsentBroker: PipeOptions ambiguous between System.IO.Pipes and
  System.IO.Pipelines (both used in the file)
- ConsentPrompt App.OnStartup: NamedPipeClientStream.ConnectAsync takes
  millis as int, not TimeSpan
- App.ViewModel renamed from App.Main to avoid colliding with WPF's
  auto-generated static void Main entry point
- AgentConnection missing using System.IO for EndOfStreamException

Adds scripts/test.ps1 — PowerShell harness that checks the SDK, builds
the solution, and optionally launches agent + console + consent prompt
for a 127.0.0.1 loopback test on Windows.

+++ src/TechSupport.Agent/Capture/DxgiScreenCapture.cs
@@ -79,8 +79,8 @@ public sealed class DxgiScreenCapture : IScreenCapture
 
         var stagingDesc = new Texture2DDescription
         {
-            Width = (uint)Width,
-            Height = (uint)Height,
+            Width = Width,
+            Height = Height,
             MipLevels = 1,
             ArraySize = 1,
             Format = Format.B8G8R8A8_UNorm,
@@ -106,7 +106,7 @@ public sealed class DxgiScreenCapture : IScreenCapture
         if (_frameAcquired) ReleaseFrame();
 
         var hr = _duplication.AcquireNextFrame(
-            (uint)timeoutMs,
+            timeoutMs,
             out _,
             out var resource);
